requirements.txt
streamlit>=1.28.0
crewai>=0.80.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
yfinance>=0.2.36
duckduckgo-search>=7.3.0
matplotlib>=3.8.0
plotly>=5.20.0
pandas>=2.2.0
python-dotenv>=1.0.0
pillow>=10.3.0